Add tests for getting event activity

+++ EventPlanning.BL/EventManager.cs
@@ -193,6 +193,11 @@ namespace EventPlanning.BL
             var activityId = GetActivityId(activityData.Title);
             var ea = repository.GetEventActivity(eventId, (int)activityId);
 
+            if (ea == null)
+            {
+                return false;
+            }
+
             return repository.Delete(ea);
         }
 
@@ -239,9 +244,25 @@ namespace EventPlanning.BL
             return false;
         }
 
+        /// <summary>
+        /// Returns event by event Id
+        /// </summary>
+        /// <param name="eventId"></param>
+        /// <returns></returns>
         public Event GetEvent(int eventId)
         {
             return repository.GetEvent(eventId);
+        }
+
+        /// <summary>
+        /// Return event activity
+        /// </summary>
+        /// <param name="eventId"></param>
+        /// <param name="activityId"></param>
+        /// <returns></returns>
+        public EventActivity GetEventActivity(int eventId, int activityId)
+        {
+            return repository.GetEventActivity(eventId, activityId);
         }
 
         /// <summary>
